fix: extract shared CSRF token utilities into csrf.ts
Create packages/web/src/lib/csrf.ts exporting getCsrfToken() and
shouldAttachCsrf(), consolidating the duplicated logic from adminFetch.ts
and api.ts. Both files now import from the shared module.

The shared version includes the SSR guard (typeof document === 'undefined')
previously only in api.ts.
